(WIP, compiles) The revolution breaks through. Fixed or commented out things so that Devil Ray compiles.
Commented out FaceElement-related code.
Commented out instantiations of Mesh for triangle/tetrahedra, until we update RefBox/splitting to use SubRef<etype>
Commented out instantiations of Mesh for surface meshes, until locate() is separated to a 3D->3D specialization of Mesh.

Haven't run or tested, just compiles.

+++ src/dray/data_set.cpp
@@ -62,9 +62,9 @@ template class DataSet<float32, MeshElem<float32, 3u, ElemType::Quad, Order::Gen
 template class DataSet<float32, MeshElem<float32, 2u, ElemType::Tri, Order::General>>;
 template class DataSet<float32, MeshElem<float32, 3u, ElemType::Tri, Order::General>>;
 
-template class DataSet<float32, MeshElem<float64, 2u, ElemType::Quad, Order::General>>;
-template class DataSet<float32, MeshElem<float64, 3u, ElemType::Quad, Order::General>>;
-template class DataSet<float32, MeshElem<float64, 2u, ElemType::Tri, Order::General>>;
-template class DataSet<float32, MeshElem<float64, 3u, ElemType::Tri, Order::General>>;
+template class DataSet<float64, MeshElem<float64, 2u, ElemType::Quad, Order::General>>;
+template class DataSet<float64, MeshElem<float64, 3u, ElemType::Quad, Order::General>>;
+template class DataSet<float64, MeshElem<float64, 2u, ElemType::Tri, Order::General>>;
+template class DataSet<float64, MeshElem<float64, 3u, ElemType::Tri, Order::General>>;
 
 } // namespace dray
